Replace https://oauth2.dance with Github pages.
https://oauth2.dance hosted some static content that reflected the OAuth2
token back to you.  After some research, it appears that this domain was
personally owned and hosted by junyer, who sadly passed away last year.
The domain subsequently expired and was reregistered by a third party
and is currently not resolving.

This is my proposal to fix this: host the required HTML via Github pages
from this repo itself.  In order to work, you will need to create a
"gh-pages" branch containing (at least) the oauth2.dance.html file.